Added Concern for session Users
