+++ app/controllers/concerns/encoded_token_concern.rb
@@ -1,0 +1,6 @@
+module EncodedTokenConcern
+    extend ActiveSupport::Concern
+    def encode_token(payload)
+        JWT.encode(payload, "my_secret")
+    end
+end
